Top-level class issues fixed
