+++ src/ch8/ByTwos.java
@@ -1,0 +1,40 @@
+package ch8;
+
+class ByTwos implements Series {
+
+    private int start, val, prev;
+
+    ByTwos() {
+        start = 0;
+        val = 0;
+        prev = -2;
+    }
+
+    @Override
+    public int getNext() {
+        prev = val;
+        val += 2;
+        return val;
+
+    }
+
+    @Override
+    public void reset() {
+        start = 0;
+        val = 0;
+        prev = -2;
+    }
+
+    @Override
+    public void setStart(int x) {
+        start = x;
+        val = x;
+        prev = x-2;
+
+    }
+
+    int getPrevious(){
+        return prev;
+    }
+
+}
