+++ src/ch8/SeriesDemo.java
@@ -1,80 +1,5 @@
 package ch8;
 
-class ByTwos implements Series {
-
-    private int start, val, prev;
-
-    ByTwos() {
-        start = 0;
-        val = 0;
-        prev = -2;
-    }
-
-    @Override
-    public int getNext() {
-        prev = val;
-        val += 2;
-        return val;
-
-    }
-
-    @Override
-    public void reset() {
-        start = 0;
-        val = 0;
-        prev = -2;
-    }
-
-    @Override
-    public void setStart(int x) {
-        start = x;
-        val = x;
-        prev = x-2;
-
-    }
-
-    int getPrevious(){
-        return prev;
-    }
-
-}
-
-class ByThrees implements Series {
-
-    private int start, val, prev;
-
-    ByThrees() {
-        start = 0;
-        val = 0;
-        prev = -3;
-    }
-
-    @Override
-    public int getNext() {
-        prev = val;
-        val += 3;
-        return val;
-    }
-
-    @Override
-    public void reset() {
-        start = 0;
-        val = 0;
-        prev = -3;
-    }
-
-    @Override
-    public void setStart(int x) {
-        start = x;
-        val = x;
-        prev = x-3;
-
-    }
-
-    int getPrevious(){
-        return prev;
-    }
-}
 
 public class SeriesDemo {
     public static void main(String[] args) {
